refactor: доработан код

+++ gateway/src/main/java/ru/practicum/shareit/item/dto/ItemRequestDTO.java
@@ -24,5 +24,5 @@ public class ItemRequestDTO {
     BookingRequestDTO lastBooking;
     BookingRequestDTO nextBooking;
     List<CommentRequestDTO> comments;
-    Integer requestId;
+    Long requestId;
 }
